+++ src/test/java/project/volunteer/domain/scheduleParticipation/repository/ScheduleParticipationRepositoryTest.java
@@ -8,7 +8,7 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import project.volunteer.common.RepositoryTest;
+import project.volunteer.support.RepositoryTest;
 import project.volunteer.domain.participation.domain.Participant;
 import project.volunteer.domain.recruitment.domain.Recruitment;
 import project.volunteer.domain.recruitment.domain.VolunteerType;
