+++ src/test/java/project/volunteer/support/RepositoryTest.java
@@ -1,4 +1,4 @@
-package project.volunteer.common;
+package project.volunteer.support;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
